use cents subunit for money

+++ lib/money-rails/configuration.rb
@@ -75,7 +75,7 @@ module MoneyRails
 
     # Default ActiveRecord migration configuration values for columns
     mattr_accessor :amount_column
-    @@amount_column = { postfix: '_cents', type: :integer, null: false, default: 0, present: true }
+    @@amount_column = { postfix: '_cents', type: :float, null: false, default: 0, present: true }
 
     mattr_accessor :currency_column
     @@currency_column = { postfix: '_currency', type: :string, null: false, default: 'USD', present: true }
